REFACTOR: moved query builder under models
